rename: change directory path about Header, Footer (#18)
