+++ strawberry/src/layout/components/footer/index.ts
