+++ strawberry/src/layout/components/header/HeaderLogo.tsx
