+++ strawberry/src/layout/components/header/HeaderMenuButtons.tsx
@@ -1,4 +1,4 @@
-import DefaultButton from "../../design_system/styles/DefaultButton";
+import DefaultButton from "../../../modules/core/design_system/styles/DefaultButton";
 import styled from "styled-components";
 import { Link } from "react-router-dom";
