+++ strawberry/src/layout/components/header/LoginButton.tsx
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import DefaultButton from "../../design_system/styles/DefaultButton";
+import DefaultButton from "../../../modules/core/design_system/styles/DefaultButton";
 import styled from "styled-components";
 
 // link 태그 추가
